fix on comments
