+++ Api/StandardResult.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Kilik\TableBundle\Api;
+
+class StandardResult extends AbstractResult
+{
+}
