+++ Resources/config/services.yml
@@ -3,4 +3,7 @@ services:
     kilik_table:
         class: Kilik\TableBundle\Services\TableService
         arguments:    ["@twig","@form.factory"]
+    kilik_table_api:
+        class: Kilik\TableBundle\Services\TableApiService
+        arguments:    ["@twig","@form.factory"]
